OQ-139: Functions added for highest/lowest
git-svn-id: svn://svn.code.sf.net/p/oquote/code@381 22cb28ec-0345-0410-b6dc-f8d001edd02c

+++ root/projects/openquote/1.1/openquote/openquote.ear/lib/openquote.jar/com/ail/openquote/Functions.java
@@ -153,7 +153,7 @@ public class Functions {
         
         for(Attribute attr: numbers) {
             if (attr.isNumberType()) {
-                if (highest==null || ((Number)attr.getObject()) > ((Number)highest.getObject())) {
+                if (highest==null || ((Number)attr.getObject()).floatValue() > ((Number)highest.getObject()).floatValue()) {
                 	highest=attr;
                 }
             }
@@ -181,7 +181,7 @@ public class Functions {
         
         for(Attribute attr: numbers) {
             if (attr.isNumberType()) {
-                if (lowest==null || ((Number)attr.getObject()) < ((Number)lowest.getObject())) {
+                if (lowest==null || ((Number)attr.getObject()).floatValue() < ((Number)lowest.getObject()).floatValue()) {
                 	lowest=attr;
                 }
             }
